Fix for ENG-3443: Fixed test_get_downtime unit failure.
1) I've updated the 'EXPECTED_ROWS' value fix the issue
2) Also fixed one static type checking error.

+++ smsdk/ma_session.py
@@ -44,7 +44,7 @@ class MaSession:
         self,
         endpoint: str,
         method: str = "get",
-        _limit: float = np.Inf,
+        _limit: int = np.Inf,
         _offset: int = 0,
         **url_params: t_.Any,
     ) -> t_.List[t_.Dict[str, t_.Any]]:
